feat(rstream): #281 update Subscription error/teardown logic
- replace old `Subscription` class w/ what was recently `Sub2` (removed)
- update/fix done(), subscribe()/unsubscribe() logic
- update related constructs (Stream, StreamSync, MetaStream, etc.)
- update Stream ctor (and factory fns) to support error handler opts arg
- update Timeout error dispatch
- fix typehints
