+++ packages/rstream/src/from/promises.ts
@@ -1,5 +1,5 @@
 import { mapcat } from "@thi.ng/transducers";
-import type { CommonOpts, ISubscription } from "../api";
+import type { ISubscription, WithErrorHandlerOpts } from "../api";
 import { optsWithID } from "../utils/idgen";
 import { fromPromise } from "./promise";
 
@@ -38,10 +38,11 @@ import { fromPromise } from "./promise";
  * ```
  *
  * @param promises -
+ * @param opts -
  */
 export const fromPromises = <T>(
     promises: Iterable<T | PromiseLike<T>>,
-    opts?: Partial<CommonOpts>
+    opts?: Partial<WithErrorHandlerOpts>
 ): ISubscription<T[], T> =>
     fromPromise(Promise.all(promises), optsWithID("promises", opts)).transform(
         mapcat((x: T[]) => x)
