+++ packages/rstream/src/metastream.ts
@@ -160,7 +160,7 @@ export class MetaStream<A, B> extends Subscription<A, B> {
     }
 
     unsubscribe(sub?: Subscription<B, any>) {
-        if (this.stream && (!sub || this.subs.length === 1)) {
+        if (this.stream && (!sub || this.subs.size === 1)) {
             this.detach(!sub);
         }
         return super.unsubscribe(sub);
