+++ packages/rstream/src/subs/resolve.ts
@@ -70,7 +70,11 @@ export class Resolver<T> extends Subscription<Promise<T>, T> {
     }
 
     done() {
-        if (this.parent!.getState() === State.DONE && this.outstanding === 0) {
+        if (
+            this.parent &&
+            this.parent.getState() === State.DONE &&
+            this.outstanding === 0
+        ) {
             super.done();
         }
     }
