+++ packages/rstream/src/subs/timeout.ts
@@ -64,18 +64,19 @@ class Timeout<T> extends Subscription<T, T> {
     reset() {
         this.timeoutId = setTimeout(() => {
             if (this.state < State.DONE) {
-                this.error(
+                this.dispatchTo(
+                    "error",
                     this.errorObj ||
                         new Error(
-                            `Timeout stream "${this.id}" after ${this.timeoutMs} ms`
+                            `Timeout '${this.id}' after ${this.timeoutMs} ms`
                         )
                 );
             }
         }, this.timeoutMs);
     }
 
-    cleanup(): void {
+    release() {
         clearTimeout(this.timeoutId);
-        super.cleanup();
+        super.release();
     }
 }
